Removed blank line requirement and further prevented blank notes
A user can no longer send a note with a title and just blank lines
because all leading \n characters are now stripped from the body,
making a note with just newlines zero length.

This also has the effect of allowing the second line to blank or not
at the user's preference. All leading blank lines will be stripped
so that the note begins with a content line once uploaded.
Signed-off-by: Jonas Buckner <buckner.jonas@gmail.com>